backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
orjson==3.10.7
pydantic-settings==2.5.2
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.4
celery[redis]==5.4.0
flower==2.0.1
tenacity==8.3.0
httpx==0.27.2
requests==2.31.0
python-dateutil==2.8.2
google-api-python-client==2.145.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
supabase==2.4.1
aiofiles==23.2.1
itsdangerous==2.2.0
passlib[bcrypt]==1.7.4
pyjwt[crypto]==2.9.0
sse-starlette==1.8.2
python-telegram-bot==21.6

